MISUV-8167 Move the createOptOutProposition factory method to the OptOutProposition object
This makes these related elements more cohesive and will allow construction of the OptOutProposition from the repository more easily
